+++ frontend/v1/domain/entities/record/Record.ts
@@ -18,6 +18,7 @@ export class Record {
     public readonly fields: Field[],
     public answer: RecordAnswer,
     private readonly suggestions: Suggestion[],
+    public readonly score: number,
     public readonly page: number
   ) {
     this.completeQuestion();